+++ docs/CHANGELOG.rst
@@ -44,8 +44,8 @@ Added
 - Add FastQC option ``nogroup`` to ``bbduk-single`` and ``bbduk-paired`` processes
 - Add CPM normalization to ``htseq-count-raw`` process
 - Add ``workflow-bbduk-star-htseq-paired``
-- Add ``prepare-annotation-dexseq`` and
-  ``differentialexpression-dexseq`` processes
+- Add ``prepare-annotation-dexseq``, ``differentialexpression-dexseq``,
+  and ``plot-dexeq`` processes
 
 Fixed
 -----